+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1434,7 +1434,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsDatasetsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-02-18T05:54:05.901Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-02-18T08:50:52.956Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1502,7 +1502,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-02-18T05:54:05.900Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-02-18T08:50:52.955Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1570,7 +1570,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-02-18T05:54:05.899Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-02-18T08:50:52.954Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1665,7 +1665,7 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-18T05:53:50.141Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-18T08:50:37.645Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")