+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHQN05V03WR0Q5BWP7KJGXHQ"
+        "01KHQZ3W9N5MQY5CFVEMHF8D10"
     )
     r"""The unique identifier of the data source"""